improve package structure and move the viewmodel to be used by the screen instead of the activity
