+++ app/src/main/java/com/maderskitech/androidcodingexerciseapp/application/di/AppModule.kt
@@ -1,9 +1,9 @@
 package com.maderskitech.androidcodingexerciseapp.application.di
 
-import com.maderskitech.androidcodingexerciseapp.presentation.MainViewModel
+import com.maderskitech.androidcodingexerciseapp.presentation.displayitems.DisplayItemsViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val appModule = module {
-    viewModelOf(::MainViewModel)
+    viewModelOf(::DisplayItemsViewModel)
 }
