+++ app/src/main/java/com/maderskitech/androidcodingexerciseapp/presentation/displayitems/DisplayItemsViewModel.kt
@@ -1,4 +1,4 @@
-package com.maderskitech.androidcodingexerciseapp.presentation
+package com.maderskitech.androidcodingexerciseapp.presentation.displayitems
 
 import android.util.Log
 import androidx.compose.runtime.getValue
@@ -6,16 +6,20 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.maderskitech.androidcodingexerciseapp.presentation.model.ItemGroup
+import com.maderskitech.androidcodingexerciseapp.presentation.displayitems.model.ItemGroup
 import com.maderskitech.kmpcodingexercisenetwork.domain.respository.ItemRepository
 import kotlinx.coroutines.launch
 
-class MainViewModel(private val itemsRepository: ItemRepository) : ViewModel() {
+class DisplayItemsViewModel(private val itemsRepository: ItemRepository) : ViewModel() {
     var isLoading by mutableStateOf(false)
         private set
 
     var itemGroups by mutableStateOf<List<ItemGroup>>(emptyList())
         private set
+
+    init {
+        fetchItems()
+    }
 
     fun fetchItems() {
         isLoading = true
